Consolidate actor/item counter click handlers
